Create code1.9
We have started our code with the top-level function main(); it is required and special in nature because this is how the application executes. So, inside the main() function, we have called a function doSomething(), which in turn calls the lifeIsShort() function. Each function gives a display output with print(); this is a handy way to display any output. We have covered many things in our first program.